pkg/typeconv/proto.go
package typeconv

import (
	"errors"
	"fmt"

	"google.golang.org/protobuf/proto"
)

func FromProto[T any](bs []byte) (*T, error) {
	if len(bs) == 0 {
		return nil, errors.New("protobuf data is empty")
	}
	var t T
	msg, ok := any(&t).(proto.Message)
	if !ok {
		return nil, fmt.Errorf("type %T does not implement proto.Message", t)
	}
	err := proto.Unmarshal(bs, msg)
	if err != nil {
		return nil, err
	}
	return &t, nil
}

func MustFromProto[T any](bs []byte) *T {
	var t T
	msg, ok := any(&t).(proto.Message)
	if !ok {
		panic(fmt.Sprintf("type *%T does not implement proto.Message", t))
	}
	err := proto.Unmarshal(bs, msg)
	if err != nil {
		panic(fmt.Sprintf("unmarshal protobuf data to *T(type=%T) error: %v", t, err))
	}
	return &t
}
